FIR/UAST: commonize abstraction of Kotlin UClass
